Cleaning up the branch to re write multitasking again.
