+++ MWP/lib/core/class.lua
@@ -1,96 +1,88 @@
 local tableClone = nym.tableClone
 
 function Class(attributes)
-    -- @FUNCTION CLASS @PARAMS {metatable, constructor, extends, implements [, ...]}
     local new_class = {}
-    local inst_mt
+    local inst_mt = {}
+    local class_mt = {}
 
-    local function lookupInClass(val)
-        -- If an instance of a class needs to lookup a value, let's make sure we
-        -- don't modify any attributes of the class, but rather duplicate the
-        -- attribute.
-        if (type(val) == 'table') then
-            print('Preparing to clone table')
-            return tableClone(val)
-        else
-            return val
+    function new_class.lookupInClass(class,key)
+        local val = try {
+            body = function()
+                local val = class[key]
+                if (type(val) == 'table') then
+                    return tableClone(val)
+                else
+                    return val
+                end
+            end,
+            catch = function(ex)
+                ex:changeType('ClassLookupException')
+                ex:throw()
+            end
+        }
+        return val
+    end
+
+    local function implementInterface(interface)
+        for k, attr in pairs(interface) do
+            if string.sub(k,1,2) ~= '__' then
+                -- If the interface is a module, we ignore module metadata
+                local val = new_class[k]
+                if val == nil or type(attr) ~= type(val) then
+                    local ex_msg = 'Class failed to implement ' .. tostring(k) .. '.'
+                    local ex = Exception:new(ex_msg,'ClassInterfaceException')
+                    ex:throw(0)
+                end
+            end
         end
     end
 
-    for key, attribute in pairs(attributes) do
-        new_class[key] = attribute -- For any methods defined in the parameters of Class, add them to our class that we are creating.
+    -- Now we add the attributes to the class
+    for k,attr in pairs(attributes) do
+        new_class[k] = attr
     end
 
     if attributes.implements then
         if #attributes.implements > 0 then
-            for _, interface in pairs(attributes.implements) do
-                for key, attribute in pairs(interface) do
-                    if string.sub(key,1,2) ~= '__' then -- If the interface is a module, we want to ignore module metadata.
-                        if new_class[key] == nil or type(attribute) ~= type(new_class[key]) then
-                            error('Class failed to implement ' .. key .. '.',2) -- Raise it up an env to the caller.
-                        end
+            for _,interface in pairs(attributes.implements) do
+                try {
+                    body = implementInterface,
+                    args = interface,
+                    catch = function(ex)
+                        ex:throw(3)
                     end
-                end
+                }
             end
         else
-            for key, attribute in pairs(attributes.implements) do -- If the table is of length 0, the implements table *is* an interface.
-                if string.sub(key,1,2) ~= '__' then -- If the interface is a module, we want to ignore module metadata.
-                    if new_class[key] == nil or type(attribute) ~= type(new_class[key]) then
-                        error('Class failed to implement ' .. key .. '.',2) -- Raise it up an env to the caller.
-                    end
+            -- In this case, the implements table is an interface itself since it is of length zero.
+            try {
+                body = implementInterface,
+                args = attributes.implements,
+                catch = function(ex)
+                    ex:throw(3)
                 end
-            end
+            }
         end
-    end
-
-    if attributes.extends then
-        if attributes.metatable then
-            class_mt = tableClone(attributes.metatable)
-            class_mt.__index = function(_class,key)
-                local ok, result = pcall(lookupInClass, attributes.extends[key])
-                if ok then
-                    _class[key] = result
-                    return _class[key]
-                else
-                    error(result,2) --Should this be 3?
-                end
-            end
-        else
-            class_mt = { __index = function(_class,key)
-                local ok, result = pcall(lookupInClass, attributes.extends[key])
-                if ok then
-                    _class[key] = result
-                    return _class[key]
-                else
-                    error(result,2)
-                end
-            end }
-        end
-        setmetatable( new_class, class_mt)
     end
 
     if attributes.metatable then
+        class_mt = tableClone(attributes.metatable)
         inst_mt = tableClone(attributes.metatable)
-        inst_mt.__index = function(object,key)
-            local ok, result = pcall(lookupInClass, new_class[key])
-            if ok then
-                object[key] = result
-                return object[key]
-            else
-                error(result,2)
-            end
+    end
+
+    inst_mt.__index = function(object,key)
+        local val = new_class:lookupInClass(key)
+        object[key] = val
+        return val -- We want to make sure that we return the instance's version of the attribute 'key' NOT the class's.
+    end
+
+    if attributes.extends then
+        -- Modify the __index metamethod of the class to look at the extended Class:
+        class_mt.__index = function(this_class,key)
+            local val = new_class.lookupInClass(attributes.extends,key)
+            this_class[key] = val
+            return val
         end
-    else
-        inst_mt = { __index = function(object,key)
-            print('Looking up in class ' .. key)
-			local ok, result = pcall(lookupInClass, new_class[key])
-            if ok then
-            	object[key] = result
-				return object[key]
-			else
-				error(result,2)
-            end
-        end }
     end
 
     new_class.new = function(...)
@@ -101,12 +93,13 @@ function Class(attributes)
             newinst = {}
         end
 
-        setmetatable( newinst, inst_mt)
+        setmetatable(newinst,inst_mt)
         return newinst
     end
 
+    setmetatable(new_class,class_mt)
     return new_class
 end
 
 _module = module.exportFunction(Class)
-return Class -- For loadfile
+return Class -- This line loads the Class API without using the Module API.
