--- MWP/lib/fuel/defaultFuelStation.lua
@@ -1,5 +1,0 @@
-local Waypoint = module.require('/MWP/lib/movement/Waypoint.lua')
-
-fuelingStation = Waypoint:new('fuelingStation', vector.new(-145,93,375))
-
-_module = fuelingStation
