+++ MWP/lib/task/CommandParser.lua
@@ -1,0 +1,17 @@
+-- This is the parser that identifies what commands to be executed when provided a string.
+
+CommandParser = Class {
+    constructor = function(self,args)
+        local obj = {
+            registered_verbs = {},
+            registered_nouns = {},
+            registered_prepositions = {}
+        }
+    end,
+
+    ParseCommand = function(command_string)
+
+    end,
+
+
+}
