--- MWP/lib/yielding/task3.doc
@@ -1,108 +1,0 @@
---[[ Welcome to the new, new task API. May your stay be pleasant and full of indices
-        Sincerely,
-            Anthony and Brevan.
-]]
-
--- import rickScript :0
-
--- From task2:
---[[
-    This is the new API to handle tasks (listeners) and taskSequences (taskHandlers). We found the previous API had some major shortcomings, including:
-    1) When a coroutine yielded within a listener, the task_co would intercept the yield before the BIOS coroutine could receive the event. This prevented APIs like GPS, GPS2, and rednet from working.
-    2) We needed to modify attributes of a parent taskHandler from within a specific listener. So we need to find a clean and simple way to pass a parent object (taskSequence) to child objects (task).
-]]
-
---[[
-    Description of control flow in handy-dandy natural language:
-
-    taskFoo: Hi, to finish my action, I need requiredPromise.
-    PTS (enclosingTaskSequence): Ok, let me see if I have a task that gives you that...
-        (if yes): here, taskBar promises registeredOutcome requiredPromise. Let me run it real quick and get back to you.
-        (if no): sorry, I need to ask someone else. You'll have to wait a bit.
-    PTS (to its own enclosingTaskSequence, PPTS): Hi, I need requiredPromise, can you give me that?
-    PPTS: Let me check..
-        [ The process repeats until requiredPromise is fulfilled.]
-]]
-
--- CLASS Promise
---[[
-    Contains the metadata of a resumeEvent along with any contingent data necessary to resume a task that requests this promise.
-    -- "The promise class may not have any methods - it may just be a table"
-
-    Attributes:
-        kind: String -- A label that describes what kind of tasks can fulfill this promise (see task.registeredOutcome).
-        dataWasAccessed: Bool -- This boolean is a flag to see if a yielding_object has used this promise before.
-        resolved: Bool -- Has this task been fulfilled by another task?
-        answerData: Table -- Stores the data of the answer
-        questionData: Table -- Store the metadata of the context in which a task requested this promise.
-        askingTask: Task -- Pointer to the task that asked for this promise.
-
-    Methods:
-		new{(askingTask, questionData)}: Promise -- Creates a new promise where {attributes} be {askingTask, questionData, kind} makes the most sense, with {attributes} pulling from any of the above properties we choose to define.
-]]
-
--- INTERFACE YieldingObject
---[[
-    This is the interface that contains the methods necessary to yield and send and receive promises.
-
-    Attributes:
-        isActive: Bool -- The status of the yielding_object
-		enclosingTaskSequence: TaskSequence -- A pointer to the taskSequence that contains this yielding_object.
-        name: String
-
-    Methods:
-        yield(requiredPromises: Promise[]): Promise[] -- Yields the yielding_object, informing the enclosingTaskSequence that this yielding_object needs requiredPromises[] where requiredPromises is a table of promises necessary to resume.
-		registerToTaskSequence(taskSequence: TaskSequence): void -- Sets self.enclosingTaskSequence.
-]]
-
--- CLASS Task IMPLEMENTS YieldingObject
---[[
-    This is the class that implements YieldingObject to run a predetermined function when the tasks conditions are satisfied in addition to receiving a promise that satisfies the task's requiredPromises.
-
-    Attributes:
-        isActive: Bool -- Determines whether this task will try to do anything if task:run(...) is called.
-		enclosingTaskSequence: TaskSequence
-        name: String
-        enabled: Bool
-        requiredPromises: Promise[] -- Contains the promises needed by this task in order to resume after yielding.
-		registeredOutcome: String -- The kind of promise this task can fulfill.
-        procedure: Function -- The function that is wrapped by self.action containing the code that is executed when task:run(...) is called.
-        action: Coroutine -- The coroutine that contains this task's procedure to be resumed/executed on self:run(...).
-        condition: Function -- The function that is evaluated to see if this task is clear to run.
-
-    Methods:
-        checkCondition(): Bool -- Checks self.condition() to see if this task is clear to run.
-        run(promises: Promise[]): Table -- Runs a predefined procedure if this task's asked promises are fulfilled and checkCondition returns true. Returns information specific to the task.
-        yield(requiredPromises: Promise[]): Promise[] -- Yields this task, informing the enclosingTaskSequence of what promises this task has asked. When this task is resumed, an array of promises is returned back to the task's action when task:run(promises) is called.
-        registerToTaskSequence(taskSequence: TaskSequence): void
-        registerOutcome(newRegisteredOutcome: String): void -- Sets self.registeredOutcome.
-        resurrect(): Bool -- Recreates the action from this task's procedure, then returns whether the resurrection was a success.
-        enable(): void -- sets self.enabled = true.
-		disable(): void -- sets self.enabled = false.
-		requestPromise{attributes} : Promise -- Wrapper for Promise:new, that keeps track of askingTask.
-
-]]
-
--- CLASS TaskSequence IMPLEMENTS YieldingObject
---[[
-    This is a class to handle the sequencing between a given set of tasks, handling yielding and providing promises
-
-    Attributes:
-        isActive: Bool
-        enclosingTaskSequence: TaskSequence
-        name: String
-        enabled: Bool
-        pendingTasks: Task[] -- The table that contains all the tasks this taskSequence will run.
-        registeredTasks: Task[] -- The table that tracks all tasks that can fulfill promises asked by tasks in taskSequence.pendingTasks.
-        askedPromises: Promise[] -- A table containing the promises that this taskSequence cannot fulfill.
-    Methods:
-        yield(requiredPromises: Promise[]): Promise[] -- Yields this taskSequence, informing the enclosingTaskSequence of what promises this taskSequence needs in order to resume its pendingTasks. When this taskSequence is resumed, an array of promises is sent to this taskSequences promises table (would this always be self.askedPromises?).
-        registerToTaskSequence(taskSequence: TaskSequence or nil): void -- Sets self.enclosingTaskSequence.
-        checkPromiseFulfillment(promise: Promise): Bool - Checks if any task in registeredTasks can fulfill promise. If so, return true. If not, return false and add promise to self.askedPromises so that self.enclosingTaskSequence can try to fulfill them.
-        run(promises: Promise[]): Table -- Takes in promises and loads them into a table of promises to be reachable by self.pendingTasks. If the sequence is enabled, it then begins/resumes execution of the action for each task in self.pendingTasks.
-        getNextTask(): Task -- Returns the next task in pendingTasks to be run.
-        queueTask(task: Task): void -- Add a task to pendingTasks.
-        unqueueTask(task: Task): void -- Remove a task from pendingTasks.
-        enable(): void -- sets self.enabled = true.
-        disable(): void -- sets self.enabled = false.
-]]
